Move domain based image classification to own file
